Add tables
